Added aws init error: "config file does not exist in working directory"

+++ cmd/aws/init.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/funcgql/cli/apollo"
 	"github.com/funcgql/cli/aws"
+	"github.com/funcgql/cli/cliio"
 	"github.com/funcgql/cli/config"
 	"github.com/funcgql/cli/go/tools"
 	"github.com/pkg/errors"
@@ -24,7 +25,19 @@ var initCmd = &cobra.Command{
 			return errors.Errorf("missing AWS configuration %s", config.ConfigFilename)
 		}
 
-		fmt.Println("🌳 Setting up AWS development environment in", cfg.GraphModulesAbsPath)
+		dir := cliio.DirOf(cfg.GraphModulesAbsPath)
+
+		if _, err := dir.Exists(); err != nil {
+			return errors.Errorf("could not find config file", err)
+		}
+
+		if bool, err := dir.Exists(); !bool && err == nil {
+			return errors.Errorf("config file does not exist in working directory🤬\nCurrent working directory: ", dir)
+		}
+
+		if bool, err := dir.Exists(); bool && err == nil {
+			fmt.Println("🌳 Setting up AWS development environment in", dir)
+		}
 
 		if err := tools.InstallAllIn(cfg.GraphModulesAbsPath); err != nil {
 			return errors.Wrap(err, "failed to install go tools")
